+++ items/src/main/scala/weco/api/items/responses/LookupItemStatus.scala
@@ -3,9 +3,8 @@ package weco.api.items.responses
 import akka.http.scaladsl.server.Route
 import com.sksamuel.elastic4s.Index
 import weco.api.items.models.DisplayItemsList
-import weco.api.items.services.ItemUpdateService
+import weco.api.items.services.{ItemUpdateService, WorkLookup}
 import weco.api.search.rest.SingleWorkDirectives
-import weco.api.stacks.services.WorkLookup
 import weco.catalogue.internal_model.identifiers.CanonicalId
 import weco.catalogue.internal_model.work.{Work, WorkState}
