+++ items/src/main/scala/weco/api/items/services/WorkLookup.scala
@@ -1,9 +1,9 @@
-package weco.api.stacks.services
+package weco.api.items.services
 
 import com.sksamuel.elastic4s.{ElasticClient, Index}
 import weco.api.search.elasticsearch.{ElasticsearchError, ElasticsearchService}
-import weco.catalogue.internal_model.identifiers.CanonicalId
 import weco.catalogue.internal_model.Implicits._
+import weco.catalogue.internal_model.identifiers.CanonicalId
 import weco.catalogue.internal_model.work.Work
 import weco.catalogue.internal_model.work.WorkState.Indexed
