+++ requests/src/test/scala/weco/api/requests/RequestingScenarioTest.scala
@@ -1,16 +1,7 @@
 package weco.api.requests
 
 import akka.http.scaladsl.model.HttpMethods.POST
-import akka.http.scaladsl.model.{
-  ContentTypes,
-  HttpEntity,
-  HttpMethods,
-  HttpRequest,
-  HttpResponse,
-  RequestEntity,
-  StatusCodes,
-  Uri
-}
+import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpMethods, HttpRequest, HttpResponse, RequestEntity, StatusCodes, Uri}
 import akka.http.scaladsl.server.Route
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import org.scalatest.GivenWhenThen
@@ -18,9 +9,8 @@ import org.scalatest.concurrent.IntegrationPatience
 import org.scalatest.featurespec.AnyFeatureSpec
 import org.scalatest.matchers.should.Matchers
 import weco.api.requests.fixtures.RequestsApiFixture
-import weco.api.requests.services.{ItemLookup, RequestsService}
+import weco.api.requests.services.{ItemLookup, RequestsService, SierraRequestsService}
 import weco.catalogue.internal_model.work.generators.ItemsGenerators
-import weco.api.stacks.services.SierraService
 import weco.api.stacks.services.memory.MemoryItemLookup
 import weco.catalogue.internal_model.identifiers.{IdState, IdentifierType}
 import weco.catalogue.internal_model.work.Item
@@ -1057,7 +1047,7 @@ class RequestingScenarioTest
     }
 
     val requestsService = new RequestsService(
-      sierraService = SierraService(client, holdLimit = holdLimit),
+      sierraService = SierraRequestsService(client, holdLimit = holdLimit),
       itemLookup = itemLookup
     )
