+++ requests/src/test/scala/weco/api/requests/fixtures/SierraServiceFixture.scala
@@ -1,8 +1,8 @@
-package weco.api.stacks.fixtures
+package weco.api.requests.fixtures
 
 import akka.http.scaladsl.model.{HttpRequest, HttpResponse, Uri}
 import weco.akka.fixtures.Akka
-import weco.api.stacks.services.SierraService
+import weco.api.requests.services.SierraRequestsService
 import weco.fixtures.TestWith
 import weco.http.client.{HttpGet, HttpPost, MemoryHttpClient}
 import weco.http.fixtures.HttpFixtures
@@ -11,16 +11,16 @@ import scala.concurrent.ExecutionContext.Implicits.global
 
 trait SierraServiceFixture extends HttpFixtures with Akka {
   def withSierraService[R](
-    responses: Seq[(HttpRequest, HttpResponse)] = Seq(),
-    holdLimit: Int = 10
-  )(testWith: TestWith[SierraService, R]): R =
+                            responses: Seq[(HttpRequest, HttpResponse)] = Seq(),
+                            holdLimit: Int = 10
+                          )(testWith: TestWith[SierraRequestsService, R]): R =
     withMaterializer { implicit mat =>
       val httpClient = new MemoryHttpClient(responses) with HttpGet
-      with HttpPost {
+        with HttpPost {
         override val baseUri: Uri = Uri("http://sierra:1234")
       }
 
-      val sierraService = SierraService(httpClient, holdLimit = holdLimit)
+      val sierraService = SierraRequestsService(httpClient, holdLimit = holdLimit)
 
       testWith(sierraService)
     }
