+++ requests/src/test/scala/weco/api/requests/services/SierraRequestsServiceTest.scala
@@ -1,4 +1,4 @@
-package weco.api.stacks.services
+package weco.api.requests.services
 
 import java.net.URI
 
@@ -7,7 +7,8 @@ import org.scalatest.EitherValues
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import weco.api.stacks.fixtures.SierraServiceFixture
+import weco.api.requests.fixtures.SierraServiceFixture
+import weco.api.requests.models.{HoldAccepted, HoldRejected}
 import weco.api.stacks.models._
 import weco.catalogue.internal_model.identifiers.IdentifierType.SierraSystemNumber
 import weco.catalogue.internal_model.identifiers.SourceIdentifier
@@ -15,7 +16,7 @@ import weco.sierra.generators.SierraIdentifierGenerators
 import weco.sierra.models.fields.SierraLocation
 import weco.sierra.models.identifiers.SierraPatronNumber
 
-class SierraServiceTest
+class SierraRequestsServiceTest
     extends AnyFunSpec
     with Matchers
     with ScalaFutures
